+++ components/settings/impl/src/main/java/com/flipperdevices/settings/impl/viewmodels/ExperimentalViewModel.kt
@@ -6,14 +6,12 @@ import com.flipperdevices.core.ui.lifecycle.LifecycleViewModel
 import com.flipperdevices.filemanager.api.navigation.FileManagerEntry
 import com.flipperdevices.metric.api.MetricApi
 import com.flipperdevices.metric.api.events.SimpleEvent
-import com.flipperdevices.settings.api.SettingsFeatureEntry
 import tangle.viewmodel.VMInject
 import javax.inject.Provider
 
 class ExperimentalViewModel @VMInject constructor(
     fileManagerEntryProvider: Provider<FileManagerEntry>,
-    metricApiProvider: Provider<MetricApi>,
-    private val settingsFeatureEntry: SettingsFeatureEntry
+    metricApiProvider: Provider<MetricApi>
 ) : LifecycleViewModel() {
     private val fileManagerEntry by fileManagerEntryProvider
     private val metricApi by metricApiProvider
@@ -21,10 +19,5 @@ class ExperimentalViewModel @VMInject constructor(
     fun onOpenFileManager(navController: NavController) {
         metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_FM)
         navController.navigate(fileManagerEntry.fileManagerDestination())
-    }
-
-    fun onOpenScreenStreaming(navController: NavController) {
-        metricApi.reportSimpleEvent(SimpleEvent.EXPERIMENTAL_OPEN_SCREEN_STREAMING)
-        navController.navigate(settingsFeatureEntry.remoteControlStub())
     }
 }