+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [FIX] New mehanism emulate in Wear OS
 - [FIX] New button texts on fap errors dialog
 - [FIX] Bold Text in github link and app action
+- [FIX] Remove Control from Options
 - [FIX] Deprecated detekt rules
 - [FIX] Fix apostrophe in text "What’s New"